backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.7.2
attrs==23.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
dj-database-url==2.1.0
Django==5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-silk==5.3.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
etext==0.0.1
frozenlist==1.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gprof2dot==2024.6.6
grpcio==1.72.0rc1
grpcio-status==1.71.0
gunicorn==21.2.0
h3==4.2.1
haversine==2.9.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflect==7.5.0
lxml==4.9.3
more-itertools==10.6.0
msgpack==1.0.8
multidict==6.0.4
ndg-httpsclient==0.5.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
phonenumbers==8.13.49
Pillow==10.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pusher==3.3.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycountryinfo==1.0.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
reportlab==4.4.0
requests==2.31.0
rsa==4.9
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
timezonefinder==6.5.8
tqdm==4.66.1
Twisted==24.11.0
txaio==23.1.1
typeguard==4.4.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
yarl==1.18.3
zope.interface==7.2
